prepare statistics
