+++ crates/polars-io/src/predicates.rs
@@ -16,7 +16,7 @@ pub trait PhysicalIoExpr: Send + Sync {
 
 #[cfg(feature = "parquet")]
 pub trait StatsEvaluator {
-    fn should_read(&self, stats: &crate::parquet::predicates::BatchStats) -> PolarsResult<bool>;
+    fn should_read(&self, stats: &BatchStats) -> PolarsResult<bool>;
 }
 
 #[cfg(feature = "parquet")]
@@ -46,4 +46,129 @@ pub(crate) fn apply_predicate(
         }
     }
     Ok(())
+}
+
+/// The statistics for a column in a Parquet file
+/// or Hive partition.
+/// they typically hold
+/// - max value
+/// - min value
+/// - null_count
+#[cfg_attr(debug_assertions, derive(Debug))]
+pub struct ColumnStats {
+    field: Field,
+    // The array may hold the null count for every row group,
+    // or for a single row group.
+    null_count: Option<Series>,
+    min_value: Option<Series>,
+    max_value: Option<Series>,
+}
+
+impl ColumnStats {
+    pub fn new(
+        field: Field,
+        null_count: Option<Series>,
+        min_value: Option<Series>,
+        max_value: Option<Series>,
+    ) -> Self {
+        Self {
+            field,
+            null_count,
+            min_value,
+            max_value,
+        }
+    }
+
+    pub fn dtype(&self) -> &DataType {
+        self.field.data_type()
+    }
+
+    pub fn null_count(&self) -> Option<usize> {
+        match self.field.data_type() {
+            #[cfg(feature = "dtype-struct")]
+            DataType::Struct(_) => None,
+            _ => {
+                let s = self.null_count.as_ref()?;
+                // if all null, there are no statistics.
+                if s.null_count() != s.len() {
+                    s.sum()
+                } else {
+                    None
+                }
+            },
+        }
+    }
+
+    pub fn to_min_max(&self) -> Option<Series> {
+        let max_val = self.max_value.as_ref()?;
+        let min_val = self.min_value.as_ref()?;
+
+        let dtype = min_val.dtype();
+
+        if Self::use_min_max(dtype) {
+            let mut min_max_values = min_val.clone();
+            min_max_values.append(max_val).unwrap();
+            if min_max_values.null_count() > 0 {
+                None
+            } else {
+                Some(min_max_values)
+            }
+        } else {
+            None
+        }
+    }
+
+    pub fn to_min(&self) -> Option<&Series> {
+        let min_val = self.min_value.as_ref()?;
+        let dtype = min_val.dtype();
+
+        if !Self::use_min_max(dtype) || min_val.len() != 1 {
+            return None;
+        }
+
+        if min_val.null_count() > 0 {
+            None
+        } else {
+            Some(min_val)
+        }
+    }
+
+    pub fn to_max(&self) -> Option<&Series> {
+        let max_val = self.max_value.as_ref()?;
+        let dtype = max_val.dtype();
+
+        if !Self::use_min_max(dtype) || max_val.len() != 1 {
+            return None;
+        }
+
+        if max_val.null_count() > 0 {
+            None
+        } else {
+            Some(max_val)
+        }
+    }
+
+    fn use_min_max(dtype: &DataType) -> bool {
+        dtype.is_numeric() || matches!(dtype, DataType::Utf8 | DataType::Binary)
+    }
+}
+
+/// A collection of column stats with a known schema.
+pub struct BatchStats {
+    schema: Schema,
+    stats: Vec<ColumnStats>,
+}
+
+impl BatchStats {
+    pub fn new(schema: Schema, stats: Vec<ColumnStats>) -> Self {
+        Self { schema, stats }
+    }
+
+    pub fn get_stats(&self, column: &str) -> polars_core::error::PolarsResult<&ColumnStats> {
+        self.schema.try_index_of(column).map(|i| &self.stats[i])
+    }
+
+    pub fn schema(&self) -> &Schema {
+        &self.schema
+    }
 }
